Generalize orchard_zsa for backward bompatibility with non-ZSA functionality
- Introduced OrchardDomain trait in note_encryption.rs to abstract differences between Orchard and Orchard ZSA.
- Implemented Domain for new generic struct OrchardType<D: OrchardDomain>.
- Added Rust macro define_note_byte_types to simplify type definitions for OrchardDomain.
- Modified TransmittedNoteCiphertext struct to be generic, accommodating different ciphertext lengths.
- Reverted module names to note_encryption_v2.rs and note_encryption_v3.rs for clarity in representing Orchard versions.
- Updated the rest of orchard_zsa codebase and unit tests to align with these generalizations.
