+++ benches/note_decryption.rs
@@ -5,7 +5,8 @@ use orchard::{
     circuit::ProvingKey,
     keys::{FullViewingKey, PreparedIncomingViewingKey, Scope, SpendingKey},
     note::AssetBase,
-    note_encryption_orchardzsa::{CompactAction, OrchardZSADomain},
+    note_encryption::{CompactAction, OrchardType},
+    note_encryption_v3::OrchardDomainV3,
     value::NoteValue,
     Anchor, Bundle,
 };
@@ -14,6 +15,8 @@ use zcash_note_encryption_zsa::{batch, try_compact_note_decryption, try_note_dec
 
 #[cfg(unix)]
 use pprof::criterion::{Output, PProfProfiler};
+
+type OrchardV3 = OrchardType<OrchardDomainV3>;
 
 fn bench_note_decryption(c: &mut Criterion) {
     let rng = OsRng;
@@ -70,7 +73,7 @@ fn bench_note_decryption(c: &mut Criterion) {
                 None,
             )
             .unwrap();
-        let bundle: Bundle<_, i64> = builder.build(rng).unwrap();
+        let bundle: Bundle<_, i64, OrchardDomainV3> = builder.build(rng).unwrap();
         bundle
             .create_proof(&pk, rng)
             .unwrap()
@@ -79,7 +82,7 @@ fn bench_note_decryption(c: &mut Criterion) {
     };
     let action = bundle.actions().first();
 
-    let domain = OrchardZSADomain::for_action(action);
+    let domain = OrchardV3::for_action(action);
 
     let compact = {
         let mut group = c.benchmark_group("note-decryption");
@@ -120,15 +123,10 @@ fn bench_note_decryption(c: &mut Criterion) {
         let ivks = 2;
         let valid_ivks = vec![valid_ivk; ivks];
         let actions: Vec<_> = (0..100)
-            .map(|_| (OrchardZSADomain::for_action(action), action.clone()))
+            .map(|_| (OrchardV3::for_action(action), action.clone()))
             .collect();
         let compact: Vec<_> = (0..100)
-            .map(|_| {
-                (
-                    OrchardZSADomain::for_action(action),
-                    CompactAction::from(action),
-                )
-            })
+            .map(|_| (OrchardV3::for_action(action), CompactAction::from(action)))
             .collect();
 
         let mut group = c.benchmark_group("batch-note-decryption");
