+++ src/bundle/batch.rs
@@ -4,8 +4,10 @@ use rand::{CryptoRng, RngCore};
 use tracing::debug;
 
 use super::{Authorized, Bundle};
+
 use crate::{
     circuit::VerifyingKey,
+    note_encryption::OrchardDomain,
     primitives::redpallas::{self, Binding, SpendAuth},
 };
 
@@ -35,9 +37,9 @@ impl BatchValidator {
     }
 
     /// Adds the proof and RedPallas signatures from the given bundle to the validator.
-    pub fn add_bundle<V: Copy + Into<i64>>(
+    pub fn add_bundle<V: Copy + Into<i64>, D: OrchardDomain>(
         &mut self,
-        bundle: &Bundle<Authorized, V>,
+        bundle: &Bundle<Authorized, V, D>,
         sighash: [u8; 32],
     ) {
         for action in bundle.actions().iter() {
