+++ src/note_encryption_v3.rs
@@ -1,0 +1,208 @@
+//! In-band secret distribution for Orchard bundles.
+
+use zcash_note_encryption_zsa::{AEAD_TAG_SIZE, MEMO_SIZE};
+
+use crate::{
+    note_encryption::{
+        build_base_note_plaintext_bytes, define_note_byte_types, Memo, OrchardDomain,
+        COMPACT_NOTE_SIZE_V2, COMPACT_NOTE_SIZE_V3,
+    },
+    Note,
+};
+
+define_note_byte_types!(COMPACT_NOTE_SIZE_V3);
+
+/// FIXME: add doc
+#[derive(Debug, Clone)]
+pub struct OrchardDomainV3;
+
+impl OrchardDomain for OrchardDomainV3 {
+    const COMPACT_NOTE_SIZE: usize = COMPACT_NOTE_SIZE;
+
+    type NotePlaintextBytes = NotePlaintextBytes;
+    type NoteCiphertextBytes = NoteCiphertextBytes;
+    type CompactNotePlaintextBytes = CompactNotePlaintextBytes;
+    type CompactNoteCiphertextBytes = CompactNoteCiphertextBytes;
+
+    fn build_note_plaintext_bytes(note: &Note, memo: &Memo) -> Self::NotePlaintextBytes {
+        let mut np = build_base_note_plaintext_bytes(0x03, note);
+
+        np[COMPACT_NOTE_SIZE_V2..COMPACT_NOTE_SIZE_V3].copy_from_slice(&note.asset().to_bytes());
+        np[COMPACT_NOTE_SIZE_V3..].copy_from_slice(memo);
+
+        NotePlaintextBytes(np)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use proptest::prelude::*;
+    use rand::rngs::OsRng;
+    use zcash_note_encryption_zsa::{
+        try_compact_note_decryption, try_note_decryption, try_output_recovery_with_ovk, Domain,
+        EphemeralKeyBytes,
+    };
+
+    use super::{NoteCiphertextBytes, OrchardDomainV3};
+    use crate::{
+        action::Action,
+        keys::{
+            DiversifiedTransmissionKey, Diversifier, EphemeralSecretKey, IncomingViewingKey,
+            OutgoingViewingKey, PreparedIncomingViewingKey,
+        },
+        note::{
+            testing::arb_note, AssetBase, ExtractedNoteCommitment, Nullifier, RandomSeed,
+            TransmittedNoteCiphertext,
+        },
+        note_encryption::{note_version, prf_ock_orchard, CompactAction, OrchardType},
+        primitives::redpallas,
+        value::{NoteValue, ValueCommitment},
+        Address, Note,
+    };
+
+    type OrchardV3 = OrchardType<OrchardDomainV3>;
+
+    /// Implementation of in-band secret distribution for Orchard bundles.
+    pub type OrchardNoteEncryptionV3 = zcash_note_encryption_zsa::NoteEncryption<OrchardV3>;
+
+    proptest! {
+        #[test]
+        fn test_encoding_roundtrip(
+            note in arb_note(NoteValue::from_raw(100)),
+        ) {
+            let memo = &crate::test_vectors::note_encryption::test_vectors()[0].memo;
+
+            // Encode.
+            let mut plaintext = OrchardV3::note_plaintext_bytes(&note, memo);
+
+            // Decode.
+            let domain = OrchardV3::for_nullifier(note.rho());
+            let parsed_version = note_version(plaintext.as_mut()).unwrap();
+            let (compact,parsed_memo) = domain.extract_memo(&plaintext);
+
+            let (parsed_note, parsed_recipient) = OrchardV3::orchard_parse_note_plaintext_without_memo(&domain, &compact,
+                |diversifier| {
+                    assert_eq!(diversifier, &note.recipient().diversifier());
+                    Some(*note.recipient().pk_d())
+                }
+            ).expect("Plaintext parsing failed");
+
+            // Check.
+            assert_eq!(parsed_note, note);
+            assert_eq!(parsed_recipient, note.recipient());
+            assert_eq!(&parsed_memo, memo);
+            assert_eq!(parsed_version, 0x03);
+        }
+    }
+
+    #[test]
+    fn test_vectors() {
+        let test_vectors = crate::test_vectors::note_encryption_v3::test_vectors();
+
+        for tv in test_vectors {
+            //
+            // Load the test vector components
+            //
+
+            // Recipient key material
+            let ivk = PreparedIncomingViewingKey::new(
+                &IncomingViewingKey::from_bytes(&tv.incoming_viewing_key).unwrap(),
+            );
+            let ovk = OutgoingViewingKey::from(tv.ovk);
+            let d = Diversifier::from_bytes(tv.default_d);
+            let pk_d = DiversifiedTransmissionKey::from_bytes(&tv.default_pk_d).unwrap();
+
+            // Received Action
+            let cv_net = ValueCommitment::from_bytes(&tv.cv_net).unwrap();
+            let rho = Nullifier::from_bytes(&tv.rho).unwrap();
+            let cmx = ExtractedNoteCommitment::from_bytes(&tv.cmx).unwrap();
+
+            let esk = EphemeralSecretKey::from_bytes(&tv.esk).unwrap();
+            let ephemeral_key = EphemeralKeyBytes(tv.ephemeral_key);
+
+            // Details about the expected note
+            let value = NoteValue::from_raw(tv.v);
+            let rseed = RandomSeed::from_bytes(tv.rseed, &rho).unwrap();
+
+            //
+            // Test the individual components
+            //
+
+            let shared_secret = esk.agree(&pk_d);
+            assert_eq!(shared_secret.to_bytes(), tv.shared_secret);
+
+            let k_enc = shared_secret.kdf_orchard(&ephemeral_key);
+            assert_eq!(k_enc.as_bytes(), tv.k_enc);
+
+            let ock = prf_ock_orchard(&ovk, &cv_net, &cmx.to_bytes(), &ephemeral_key);
+            assert_eq!(ock.as_ref(), tv.ock);
+
+            let recipient = Address::from_parts(d, pk_d);
+
+            let asset = AssetBase::from_bytes(&tv.asset).unwrap();
+
+            let note = Note::from_parts(recipient, value, asset, rho, rseed).unwrap();
+            assert_eq!(ExtractedNoteCommitment::from(note.commitment()), cmx);
+
+            let action = Action::from_parts(
+                // rho is the nullifier in the receiving Action.
+                rho,
+                // We don't need a valid rk for this test.
+                redpallas::VerificationKey::dummy(),
+                cmx,
+                TransmittedNoteCiphertext {
+                    epk_bytes: ephemeral_key.0,
+                    enc_ciphertext: NoteCiphertextBytes(tv.c_enc),
+                    out_ciphertext: tv.c_out,
+                },
+                cv_net.clone(),
+                (),
+            );
+
+            //
+            // Test decryption
+            // (Tested first because it only requires immutable references.)
+            //
+
+            let domain = OrchardV3::for_nullifier(rho);
+
+            match try_note_decryption(&domain, &ivk, &action) {
+                Some((decrypted_note, decrypted_to, decrypted_memo)) => {
+                    assert_eq!(decrypted_note, note);
+                    assert_eq!(decrypted_to, recipient);
+                    assert_eq!(&decrypted_memo[..], &tv.memo[..]);
+                }
+                None => panic!("Note decryption failed"),
+            }
+
+            match try_compact_note_decryption(&domain, &ivk, &CompactAction::from(&action)) {
+                Some((decrypted_note, decrypted_to)) => {
+                    assert_eq!(decrypted_note, note);
+                    assert_eq!(decrypted_to, recipient);
+                }
+                None => panic!("Compact note decryption failed"),
+            }
+
+            match try_output_recovery_with_ovk(&domain, &ovk, &action, &cv_net, &tv.c_out) {
+                Some((decrypted_note, decrypted_to, decrypted_memo)) => {
+                    assert_eq!(decrypted_note, note);
+                    assert_eq!(decrypted_to, recipient);
+                    assert_eq!(&decrypted_memo[..], &tv.memo[..]);
+                }
+                None => panic!("Output recovery failed"),
+            }
+
+            //
+            // Test encryption
+            //
+
+            let ne = OrchardNoteEncryptionV3::new_with_esk(esk, Some(ovk), note, tv.memo);
+
+            assert_eq!(ne.encrypt_note_plaintext().as_ref(), &tv.c_enc[..]);
+            assert_eq!(
+                &ne.encrypt_outgoing_plaintext(&cv_net, &cmx, &mut OsRng)[..],
+                &tv.c_out[..]
+            );
+        }
+    }
+}
