+++ tests/builder.rs
@@ -6,7 +6,8 @@ use orchard::{
     circuit::{ProvingKey, VerifyingKey},
     keys::{FullViewingKey, PreparedIncomingViewingKey, Scope, SpendAuthorizingKey, SpendingKey},
     note::{AssetBase, ExtractedNoteCommitment},
-    note_encryption_orchardzsa::OrchardZSADomain,
+    note_encryption::OrchardType,
+    note_encryption_v3::OrchardDomainV3,
     tree::{MerkleHashOrchard, MerklePath},
     value::NoteValue,
     Anchor, Bundle, Note,
@@ -14,7 +15,13 @@ use orchard::{
 use rand::rngs::OsRng;
 use zcash_note_encryption_zsa::try_note_decryption;
 
-pub fn verify_bundle(bundle: &Bundle<Authorized, i64>, vk: &VerifyingKey, verify_proof: bool) {
+type OrchardV3 = OrchardType<OrchardDomainV3>;
+
+pub fn verify_bundle(
+    bundle: &Bundle<Authorized, i64, OrchardDomainV3>,
+    vk: &VerifyingKey,
+    verify_proof: bool,
+) {
     if verify_proof {
         assert!(matches!(bundle.verify_proof(vk), Ok(())));
     }
@@ -58,7 +65,7 @@ fn bundle_chain() {
     let recipient = fvk.address_at(0u32, Scope::External);
 
     // Create a shielding bundle.
-    let shielding_bundle: Bundle<_, i64> = {
+    let shielding_bundle: Bundle<_, i64, OrchardDomainV3> = {
         // Use the empty tree.
         let anchor = MerkleHashOrchard::empty_root(32.into()).into();
 
@@ -83,13 +90,13 @@ fn bundle_chain() {
     verify_bundle(&shielding_bundle, &vk, true);
 
     // Create a shielded bundle spending the previous output.
-    let shielded_bundle: Bundle<_, i64> = {
+    let shielded_bundle: Bundle<_, i64, OrchardDomainV3> = {
         let ivk = PreparedIncomingViewingKey::new(&fvk.to_ivk(Scope::External));
         let (note, _, _) = shielding_bundle
             .actions()
             .iter()
             .find_map(|action| {
-                let domain = OrchardZSADomain::for_action(action);
+                let domain = OrchardV3::for_action(action);
                 try_note_decryption(&domain, &ivk, action)
             })
             .unwrap();
